refactor: merge `dom.spec` and `domEvent.spec`
